+++ bot_run.py
@@ -1,0 +1,3 @@
+import discord_related
+
+discord_related.bot.run('INSERT_API_KEY_HERE')  #  'INSERT_API_KEY_HERE'   for copy-pasting when git pulling
