+++ general.py
